+++ CHANGES.rst
@@ -6,6 +6,7 @@ Terkin Datalogger CHANGES
 Development
 ===========
 - Add minimal weight scale adjustment wizard
+- Add Sequans LTE Modem support for Pycom GPy/FiPy devices. Thanks, @wetterfrosch!
 
 2020-05-04 0.9.0
 ================